Added more structure to the repository
