+++ HAR Version 1.0/data/activity1.txt
@@ -1,0 +1,1 @@
+WALKING
