+++ HAR Version 1.0/data/activity2.txt
@@ -1,0 +1,1 @@
+WALKING
