+++ HAR Version 1.0/data/position1.txt
@@ -1,0 +1,1 @@
+-265.61, -570.12
