+++ HAR Version 1.0/data/position2.txt
@@ -1,0 +1,1 @@
+246.02, 386.09
